Mechanical refactor: Move build_package_new.py to build_package.py
